+++ packages/authentication/modules/user/useUpdateUser/types.ts
@@ -1,12 +1,12 @@
-import { UseMutationOptions } from '@tanstack/react-query'
+import type { UseMutationOptions } from '@tanstack/react-query'
 
 import UserApi from '../../../services/user'
-import { CustomCookieNames } from '../../../types/auth'
-import { User, UserUpdateParams } from '../../../types/user'
+import type { CustomJWTKeyNames } from '../../../types/auth'
+import type { User, UserUpdateParams } from '../../../types/user'
 
 type ApiClass = Pick<typeof UserApi, 'updateUser'>
 
-export interface UseUpdateUserOptions<TUser extends Partial<User>> extends CustomCookieNames {
+export interface UseUpdateUserOptions<TUser extends Partial<User>> extends CustomJWTKeyNames {
   options?: UseMutationOptions<TUser, any, UserUpdateParams<TUser>, { previousUser?: TUser }>
   ApiClass?: ApiClass
 }