+++ packages/authentication/types/auth.ts
@@ -1,4 +1,4 @@
-import { MfaMethod } from './mfa'
+import type { MfaMethod } from './mfa'
 
 export interface LoginRequest {
   email: string
@@ -15,10 +15,6 @@ export interface LoginMfaResponse {
   method: MfaMethod
 }
 
-export interface LoginSimpleTokenResponse {
-  token: string
-}
-
 export interface LoginJWTResponse {
   access: string
   refresh: string
@@ -30,7 +26,6 @@ export interface LoginChangeExpiredPasswordRedirectResponse {
 
 export type LoginResponse =
   | LoginMfaResponse
-  | LoginSimpleTokenResponse
   | LoginJWTResponse
   | LoginChangeExpiredPasswordRedirectResponse
 
@@ -49,16 +44,10 @@ export interface RegisterRequest {
   password: string
 }
 
-export interface CustomCookieNames {
-  cookieName?: string
-  refreshCookieName?: string
+export interface CustomJWTKeyNames {
+  accessKeyName?: string
+  refreshKeyName?: string
 }
-
-export interface PreAuthRequest {
-  token: string
-}
-
-export type PreAuthResponse = LoginSimpleTokenResponse | LoginJWTResponse
 
 export interface ChangeExpiredPasswordRequest {
   currentPassword: string