+++ packages/core/src/form/index.tsx
@@ -1,3 +1,0 @@
-export * from './withController'
-export * from './withController/types.d'
-export * from './utils'