+++ packages/test/utils/mockFetchError/types.ts
@@ -1,0 +1,7 @@
+type MockFetchOptions = {
+  status?: Response['status']
+  method?: RequestInit['method']
+  error?: string
+}
+
+export type MockFetchError = (url: string, options?: MockFetchOptions) => void